Small modifications and rewrite README.md
