+++ README.md
@@ -1,26 +1,22 @@
-# Password Generator C++
-using only C++ to generate password according to user choice 
+# C++ Password Generator
 
+C++ simple password generator fully customizable and flexible.
 
-## Instruction to view this project 
-1) make sure you have g++ as your compiler in your mac or window 
-2) clone the project to your pc or mac to your favourite directory by copying the code snippet below 
-```
-git clone https://github.com/PaulCompSci/Random-Password-Generator-C-.git
-```
-3) copy the following command into your terminal
-```
-g++ main.cpp function/random_char.cpp function/password.cpp function/outputFunction.cpp  -o main
-```
+## How to Use
 
-4)copy the following command into your terminal
- ```
- ./main
- ```
- 
- ## Additional resources
-Want to know the strength of the password generated ? Check out [Password strength tester](https://www.passwordmonster.com) 
+1. Make sure you have g++ (on Linux/macOS) or MSVC (on Windows).
+   To get g++ on Ubuntu Linux, you can use the following command: ```sudo apt-get install g++```. If you're on Windows, download Visual Studio, place it in a new project and build/debug the program.
 
+2. Clone the repository to your computer by following the following command: ```git clone https://github.com/PaulCompSci/Random-Password-Generator.git```
 
-## My project Vs Other Project
-my project allow user to have the freedom to fully customize the password they want
+3. To compile the program (on Linux), use the following command: ```g++ main.cpp -o [whatever]```
+
+4. Run the program by executing this in your terminal: ```./[whatever]```
+
+## Additional resources
+
+Want to know the strength of the generated password? Check out this [password strength tester](https://www.passwordmonster.com)!
+
+## My project vs. Other projects
+
+My project allows the user to have full freedom over the password they want to get.
